+++ src/api/auth/auth.service.ts
@@ -1,5 +1,4 @@
 import api from '@/lib/axios';
-import { InternalAxiosRequestConfig } from 'axios';
 import {
   ChangePasswordRequest,
   ChangePasswordResponse,
@@ -11,16 +10,7 @@ const PATH = '/auth';
 
 class AuthService {
   async login(body: LoginRequest) {
-    const result = await api.post<LoginResponse>(`${PATH}/login`, body);
-    const { accessToken } = result.data;
-    if (accessToken) {
-      const requestInterceptor = (config: InternalAxiosRequestConfig) => {
-        config.headers['Authorization'] = `Bearer ${accessToken}`;
-        return config;
-      };
-      api.interceptors.request.use(requestInterceptor);
-    }
-    return result;
+    return api.post<LoginResponse>(`${PATH}/login`, body);
   }
   changePassword(body: ChangePasswordRequest) {
     return api.put<ChangePasswordResponse>(`${PATH}/password`, body);
